add dicom_server fixture to allow local tests
This commit adds a fixture called `dicom_server`
that provides server details to the tests. If
the test is not executed in an appveyor VM,
the fixture will create a local http-server,
that serves from the directory specified by
the environment variable `INM_ICF_TEST_STUDIES`.

`pytest-env` has been added to the dependencies
to make environment variables accessible within
the tests