fix redirection following, factor out sudo
This commit instructs curl to follow redirections,
in order to properly download example DICOMs.

The commit also removes all sudo calls from the
scanner output creation script. Instead the
script is called via sudo from `.appveyor.yml`

In addition the commit adds a script to check
the version compatiblity of the installed
test environment against variables that hold
the versions which are used in ICF