Simplify dataset deposit
- the structure of the deposit is simplified (only two files)
- the tooling is simplified (only uncurl for depositing too now)

This is what a deposit looks like now:

```sh
❯ ls -l dl-Z03/
.rw-rw-r-- 792M mih mih 26 Apr 11:29  P000624_dicom.tar
.rw-r--r--   52 mih mih  2 May 09:57  P000624_dicom.tar.md5sum
.rw-rw-r--  145 mih mih 23 May 10:54  P000624_XDLRA--refs
.rw-rw-r-- 285k mih mih 23 May 10:54  P000624_XDLRA--repo-export
```

The first to files where the starting point, and the other two files are
added. That is all.

Such a deposit can be cloned via the equivalent of

```
datalad clone \
  'datalad-annex::?type=external&externaltype=uncurl&url=file:///tmp/icfstore/dl-Z03/P000624_{{annex_key}}&encryption=none'
```

where the `url=` part is the only variable item. Basically it is:

`<ignore>url=<store base URL>/<studyid>/<visitid>_<always constant>`

A convenience short cut can be provided via configuration item
`datalad.clone.url-substitute...`