docs: describe scripts in detail
This adds slightly reformatted docstrings, copied from respective
scripts, into the admin documentation. Getting ahead of the current
main branch, the docs content borrows updated name and docstring from
PR #33

Admittedly, this is a lot of copy-pasted text. Ideally, we could build
docstrings as part of the doc build process, but given that we plan to
keep the design stable I think it's OK to leave it at that.