docs: increase cohesion, adjust catalog mentions
This unifies some spelling (local_dicom_store, uppercase DICOM), and
adds or tweaks some opening/closing paragraphs.

Catalog mention for browser-based access gets prefixed with an "if",
because it is unlikely that it will be generated for the ICF
datasets. I leave the paragraph nonetheless, to signal the
possibility.

Catalog mention for datalad based access is removed, because the
catalogify script hardcodes the (clone) URL to the ICF store URL, and
we are now assuming that the catalog generation happens on local
infrastructure.