Fix generation of inner tar paths
This uses `Path.relative_to()` and the `arcname` parameter of
`tar.gettarinfo()` to specify names for files in the archive. We keep
setting the base directory. The behaviour is similar to "tar -C" but
with that directory added.

This changes the previous approach, which only replaced one leading
path component. That caused potentially unwanted leading path
components to be included in the tar paths whenever the input
directory given to the script was a path with more than two
components. It was mostly noticeable when using absolute paths.
E.g. for `/tmp/dicomdir` our code would replace `/tmp` and keep
`dicomdir`; for `~/Documents/foo/dicomdir` it would would replace
`/home` and keep `username/Documents/foo/dicomdir`.

This should make the behaviour consistent with the following
docstring (although "base directory" can be slightly ambiguous):

  The input base directory itself is not put into the archive (i.e.,
  its own name is irrelevant). Instead, a top-level directory with the
  name '<study-id>_<visit_id>' is used to place all archive content
  in.