Unify -o long option to be --store-dir
This changes make_studyvisit_archive to use -o / --store-dir,
consistent with the other three scripts. The --output-dir argument is
marked as deprecated but remains supported for the time being.

Use of both long forms is guarded againist by using a mutually
exclusive group. A deprecation warning is issued when --output-dir is
used. When the deprecated parameter is removed, both the group and the
warning can be done away with.

As a side note, it seems that Python 3.13 will add "deprecated" to
argparse's add_argument:
https://docs.python.org/3.13/library/argparse.html#deprecated